Add MiniExcelWriteAdapterFactory
